+++ audit/options.go
@@ -7,18 +7,12 @@ import (
 	"strconv"
 	"strings"
 	"time"
-
-	"github.com/hashicorp/go-secure-stdlib/parseutil"
 )
 
 // getDefaultOptions returns options with their default values.
 func getDefaultOptions() options {
 	return options{
-		withNow:         time.Now(),
-		withFacility:    "AUTH",
-		withTag:         "vault",
-		withSocketType:  "tcp",
-		withMaxDuration: 2 * time.Second,
+		withNow: time.Now(),
 	}
 }
 
@@ -145,65 +139,6 @@ func WithPrefix(prefix string) Option {
 		if prefix != "" {
 			o.withPrefix = prefix
 		}
-
-		return nil
-	}
-}
-
-// WithFacility provides an Option to represent a 'facility' for a syslog sink.
-func WithFacility(facility string) Option {
-	return func(o *options) error {
-		facility = strings.TrimSpace(facility)
-
-		if facility != "" {
-			o.withFacility = facility
-		}
-
-		return nil
-	}
-}
-
-// WithTag provides an Option to represent a 'tag' for a syslog sink.
-func WithTag(tag string) Option {
-	return func(o *options) error {
-		tag = strings.TrimSpace(tag)
-
-		if tag != "" {
-			o.withTag = tag
-		}
-
-		return nil
-	}
-}
-
-// WithSocketType provides an Option to represent the socket type for a socket sink.
-func WithSocketType(socketType string) Option {
-	return func(o *options) error {
-		socketType = strings.TrimSpace(socketType)
-
-		if socketType != "" {
-			o.withSocketType = socketType
-		}
-
-		return nil
-	}
-}
-
-// WithMaxDuration provides an Option to represent the max duration for writing to a socket.
-func WithMaxDuration(duration string) Option {
-	return func(o *options) error {
-		duration = strings.TrimSpace(duration)
-
-		if duration == "" {
-			return nil
-		}
-
-		parsed, err := parseutil.ParseDurationSecond(duration)
-		if err != nil {
-			return err
-		}
-
-		o.withMaxDuration = parsed
 
 		return nil
 	}
