+++ internal/observability/event/sink_noop.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package audit
+package event
 
 import (
 	"context"
