+++ internal/observability/event/sink_syslog.go
@@ -1,29 +1,27 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package audit
+package event
 
 import (
 	"context"
 	"fmt"
-
-	"github.com/hashicorp/vault/internal/observability/event"
 
 	gsyslog "github.com/hashicorp/go-syslog"
 
 	"github.com/hashicorp/eventlogger"
 )
 
-// SyslogSink is a sink node which handles writing audit events to syslog.
+// SyslogSink is a sink node which handles writing events to syslog.
 type SyslogSink struct {
-	format format
-	logger gsyslog.Syslogger
+	requiredFormat string
+	logger         gsyslog.Syslogger
 }
 
 // NewSyslogSink should be used to create a new SyslogSink.
 // Accepted options: WithFacility and WithTag.
-func NewSyslogSink(format format, opt ...Option) (*SyslogSink, error) {
-	const op = "audit.NewSyslogSink"
+func NewSyslogSink(format string, opt ...Option) (*SyslogSink, error) {
+	const op = "event.NewSyslogSink"
 
 	opts, err := getOpts(opt...)
 	if err != nil {
@@ -35,12 +33,12 @@ func NewSyslogSink(format format, opt ...Option) (*SyslogSink, error) {
 		return nil, fmt.Errorf("%s: error creating syslogger: %w", op, err)
 	}
 
-	return &SyslogSink{format: format, logger: logger}, nil
+	return &SyslogSink{requiredFormat: format, logger: logger}, nil
 }
 
 // Process handles writing the event to the syslog.
 func (s *SyslogSink) Process(ctx context.Context, e *eventlogger.Event) (*eventlogger.Event, error) {
-	const op = "audit.(SyslogSink).Process"
+	const op = "event.(SyslogSink).Process"
 
 	select {
 	case <-ctx.Done():
@@ -49,12 +47,12 @@ func (s *SyslogSink) Process(ctx context.Context, e *eventlogger.Event) (*eventl
 	}
 
 	if e == nil {
-		return nil, fmt.Errorf("%s: event is nil: %w", op, event.ErrInvalidParameter)
+		return nil, fmt.Errorf("%s: event is nil: %w", op, ErrInvalidParameter)
 	}
 
-	formatted, found := e.Format(s.format.String())
+	formatted, found := e.Format(s.requiredFormat)
 	if !found {
-		return nil, fmt.Errorf("%s: unable to retrieve event formatted as %q", op, s.format)
+		return nil, fmt.Errorf("%s: unable to retrieve event formatted as %q", op, s.requiredFormat)
 	}
 
 	_, err := s.logger.Write(formatted)
